requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
pip==23.0.1
PyJWT==2.8.0
pytz==2023.3.post1
setuptools==66.1.1
sqlparse==0.4.4
